+++ sentry_streams/tests/adapters/arroyo/test_adapter.py
@@ -3,9 +3,7 @@ from unittest import mock
 import pytest
 from arroyo.backends.kafka.consumer import KafkaPayload
 from arroyo.backends.local.backend import LocalBroker
-from arroyo.backends.local.storages.memory import MemoryMessageStorage
 from arroyo.types import Partition, Topic
-from arroyo.utils.clock import MockedClock
 
 from sentry_streams.adapters.arroyo.adapter import (
     ArroyoAdapter,
@@ -14,10 +12,7 @@ from sentry_streams.adapters.arroyo.adapter import (
 )
 from sentry_streams.adapters.stream_adapter import RuntimeTranslator
 from sentry_streams.pipeline.pipeline import (
-    Filter,
-    KafkaSink,
     KafkaSource,
-    Map,
     Pipeline,
 )
 from sentry_streams.runner import iterate_edges
@@ -51,50 +46,14 @@ def test_kafka_sources() -> None:
     assert sources.get_consumer("source1") is not None
 
 
-@pytest.fixture
-def broker() -> LocalBroker[KafkaPayload]:
-    storage: MemoryMessageStorage[KafkaPayload] = MemoryMessageStorage()
-    broker = LocalBroker(storage, MockedClock())
-    broker.create_topic(Topic("logical-events"), 1)
-    broker.create_topic(Topic("transformed-events"), 1)
-    return broker
-
-
-def test_adapter(broker: LocalBroker[KafkaPayload]) -> None:
-    pipeline = Pipeline()
-    source = KafkaSource(
-        name="myinput",
-        ctx=pipeline,
-        logical_topic="logical-events",
-    )
-    decoder = Map(
-        name="decoder",
-        ctx=pipeline,
-        inputs=[source],
-        function=lambda msg: msg.decode("utf-8"),
-    )
-    filter = Filter(
-        name="myfilter", ctx=pipeline, inputs=[decoder], function=lambda msg: msg == "go_ahead"
-    )
-    map = Map(
-        name="mymap",
-        ctx=pipeline,
-        inputs=[filter],
-        function=lambda msg: msg + "_mapped",
-    )
-    _ = KafkaSink(
-        name="kafkasink",
-        ctx=pipeline,
-        inputs=[map],
-        logical_topic="transformed-events",
-    )
+def test_adapter(broker: LocalBroker[KafkaPayload], pipeline: Pipeline) -> None:
 
     adapter = ArroyoAdapter.build(
         {
             "sources_config": {},
             "sinks_config": {},
             "sources_override": {
-                "myinput": broker.get_consumer(source.logical_topic),
+                "myinput": broker.get_consumer("logical-events"),
             },
             "sinks_override": {
                 "kafkasink": broker.get_producer(),
